backend/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
firebase-admin
python-multipart
email-validator
bleach
sqlalchemy
psycopg2-binary
redis
cryptography
python-magic
pillow
reportlab
